Refactor code for easier sharing with DX12 version
