+++ Src/ScreenGrab.cpp
@@ -27,439 +27,18 @@
 
 #include "pch.h"
 
-#include <wincodec.h>
-
 #include "ScreenGrab.h"
 #include "DirectXHelpers.h"
 
 #include "dds.h"
 #include "PlatformHelpers.h"
+#include "LoaderHelpers.h"
 
 using Microsoft::WRL::ComPtr;
 using namespace DirectX;
 
 namespace
 {
-    class auto_delete_file
-    {
-    public:
-        auto_delete_file(HANDLE hFile) : m_handle(hFile) {}
-
-        auto_delete_file(const auto_delete_file&) = delete;
-        auto_delete_file& operator=(const auto_delete_file&) = delete;
-
-        ~auto_delete_file()
-        {
-            if (m_handle)
-            {
-                FILE_DISPOSITION_INFO info = {};
-                info.DeleteFile = TRUE;
-                (void)SetFileInformationByHandle(m_handle, FileDispositionInfo, &info, sizeof(info));
-            }
-        }
-
-        void clear() { m_handle = 0; }
-
-    private:
-        HANDLE m_handle;
-    };
-
-    class auto_delete_file_wic
-    {
-    public:
-        auto_delete_file_wic(ComPtr<IWICStream>& hFile, LPCWSTR szFile) : m_handle(hFile), m_filename(szFile) {}
-
-        auto_delete_file_wic(const auto_delete_file_wic&) = delete;
-        auto_delete_file_wic& operator=(const auto_delete_file_wic&) = delete;
-
-        ~auto_delete_file_wic()
-        {
-            if (m_filename)
-            {
-                m_handle.Reset();
-                DeleteFileW(m_filename);
-            }
-        }
-
-        void clear() { m_filename = 0; }
-
-    private:
-        LPCWSTR m_filename;
-        ComPtr<IWICStream>& m_handle;
-    };
-
-    //--------------------------------------------------------------------------------------
-    // Return the BPP for a particular format
-    //--------------------------------------------------------------------------------------
-    size_t BitsPerPixel(_In_ DXGI_FORMAT fmt)
-    {
-        switch (fmt)
-        {
-        case DXGI_FORMAT_R32G32B32A32_TYPELESS:
-        case DXGI_FORMAT_R32G32B32A32_FLOAT:
-        case DXGI_FORMAT_R32G32B32A32_UINT:
-        case DXGI_FORMAT_R32G32B32A32_SINT:
-            return 128;
-
-        case DXGI_FORMAT_R32G32B32_TYPELESS:
-        case DXGI_FORMAT_R32G32B32_FLOAT:
-        case DXGI_FORMAT_R32G32B32_UINT:
-        case DXGI_FORMAT_R32G32B32_SINT:
-            return 96;
-
-        case DXGI_FORMAT_R16G16B16A16_TYPELESS:
-        case DXGI_FORMAT_R16G16B16A16_FLOAT:
-        case DXGI_FORMAT_R16G16B16A16_UNORM:
-        case DXGI_FORMAT_R16G16B16A16_UINT:
-        case DXGI_FORMAT_R16G16B16A16_SNORM:
-        case DXGI_FORMAT_R16G16B16A16_SINT:
-        case DXGI_FORMAT_R32G32_TYPELESS:
-        case DXGI_FORMAT_R32G32_FLOAT:
-        case DXGI_FORMAT_R32G32_UINT:
-        case DXGI_FORMAT_R32G32_SINT:
-        case DXGI_FORMAT_R32G8X24_TYPELESS:
-        case DXGI_FORMAT_D32_FLOAT_S8X24_UINT:
-        case DXGI_FORMAT_R32_FLOAT_X8X24_TYPELESS:
-        case DXGI_FORMAT_X32_TYPELESS_G8X24_UINT:
-        case DXGI_FORMAT_Y416:
-        case DXGI_FORMAT_Y210:
-        case DXGI_FORMAT_Y216:
-            return 64;
-
-        case DXGI_FORMAT_R10G10B10A2_TYPELESS:
-        case DXGI_FORMAT_R10G10B10A2_UNORM:
-        case DXGI_FORMAT_R10G10B10A2_UINT:
-        case DXGI_FORMAT_R11G11B10_FLOAT:
-        case DXGI_FORMAT_R8G8B8A8_TYPELESS:
-        case DXGI_FORMAT_R8G8B8A8_UNORM:
-        case DXGI_FORMAT_R8G8B8A8_UNORM_SRGB:
-        case DXGI_FORMAT_R8G8B8A8_UINT:
-        case DXGI_FORMAT_R8G8B8A8_SNORM:
-        case DXGI_FORMAT_R8G8B8A8_SINT:
-        case DXGI_FORMAT_R16G16_TYPELESS:
-        case DXGI_FORMAT_R16G16_FLOAT:
-        case DXGI_FORMAT_R16G16_UNORM:
-        case DXGI_FORMAT_R16G16_UINT:
-        case DXGI_FORMAT_R16G16_SNORM:
-        case DXGI_FORMAT_R16G16_SINT:
-        case DXGI_FORMAT_R32_TYPELESS:
-        case DXGI_FORMAT_D32_FLOAT:
-        case DXGI_FORMAT_R32_FLOAT:
-        case DXGI_FORMAT_R32_UINT:
-        case DXGI_FORMAT_R32_SINT:
-        case DXGI_FORMAT_R24G8_TYPELESS:
-        case DXGI_FORMAT_D24_UNORM_S8_UINT:
-        case DXGI_FORMAT_R24_UNORM_X8_TYPELESS:
-        case DXGI_FORMAT_X24_TYPELESS_G8_UINT:
-        case DXGI_FORMAT_R9G9B9E5_SHAREDEXP:
-        case DXGI_FORMAT_R8G8_B8G8_UNORM:
-        case DXGI_FORMAT_G8R8_G8B8_UNORM:
-        case DXGI_FORMAT_B8G8R8A8_UNORM:
-        case DXGI_FORMAT_B8G8R8X8_UNORM:
-        case DXGI_FORMAT_R10G10B10_XR_BIAS_A2_UNORM:
-        case DXGI_FORMAT_B8G8R8A8_TYPELESS:
-        case DXGI_FORMAT_B8G8R8A8_UNORM_SRGB:
-        case DXGI_FORMAT_B8G8R8X8_TYPELESS:
-        case DXGI_FORMAT_B8G8R8X8_UNORM_SRGB:
-        case DXGI_FORMAT_AYUV:
-        case DXGI_FORMAT_Y410:
-        case DXGI_FORMAT_YUY2:
-            return 32;
-
-        case DXGI_FORMAT_P010:
-        case DXGI_FORMAT_P016:
-            return 24;
-
-        case DXGI_FORMAT_R8G8_TYPELESS:
-        case DXGI_FORMAT_R8G8_UNORM:
-        case DXGI_FORMAT_R8G8_UINT:
-        case DXGI_FORMAT_R8G8_SNORM:
-        case DXGI_FORMAT_R8G8_SINT:
-        case DXGI_FORMAT_R16_TYPELESS:
-        case DXGI_FORMAT_R16_FLOAT:
-        case DXGI_FORMAT_D16_UNORM:
-        case DXGI_FORMAT_R16_UNORM:
-        case DXGI_FORMAT_R16_UINT:
-        case DXGI_FORMAT_R16_SNORM:
-        case DXGI_FORMAT_R16_SINT:
-        case DXGI_FORMAT_B5G6R5_UNORM:
-        case DXGI_FORMAT_B5G5R5A1_UNORM:
-        case DXGI_FORMAT_A8P8:
-        case DXGI_FORMAT_B4G4R4A4_UNORM:
-            return 16;
-
-        case DXGI_FORMAT_NV12:
-        case DXGI_FORMAT_420_OPAQUE:
-        case DXGI_FORMAT_NV11:
-            return 12;
-
-        case DXGI_FORMAT_R8_TYPELESS:
-        case DXGI_FORMAT_R8_UNORM:
-        case DXGI_FORMAT_R8_UINT:
-        case DXGI_FORMAT_R8_SNORM:
-        case DXGI_FORMAT_R8_SINT:
-        case DXGI_FORMAT_A8_UNORM:
-        case DXGI_FORMAT_AI44:
-        case DXGI_FORMAT_IA44:
-        case DXGI_FORMAT_P8:
-            return 8;
-
-        case DXGI_FORMAT_R1_UNORM:
-            return 1;
-
-        case DXGI_FORMAT_BC1_TYPELESS:
-        case DXGI_FORMAT_BC1_UNORM:
-        case DXGI_FORMAT_BC1_UNORM_SRGB:
-        case DXGI_FORMAT_BC4_TYPELESS:
-        case DXGI_FORMAT_BC4_UNORM:
-        case DXGI_FORMAT_BC4_SNORM:
-            return 4;
-
-        case DXGI_FORMAT_BC2_TYPELESS:
-        case DXGI_FORMAT_BC2_UNORM:
-        case DXGI_FORMAT_BC2_UNORM_SRGB:
-        case DXGI_FORMAT_BC3_TYPELESS:
-        case DXGI_FORMAT_BC3_UNORM:
-        case DXGI_FORMAT_BC3_UNORM_SRGB:
-        case DXGI_FORMAT_BC5_TYPELESS:
-        case DXGI_FORMAT_BC5_UNORM:
-        case DXGI_FORMAT_BC5_SNORM:
-        case DXGI_FORMAT_BC6H_TYPELESS:
-        case DXGI_FORMAT_BC6H_UF16:
-        case DXGI_FORMAT_BC6H_SF16:
-        case DXGI_FORMAT_BC7_TYPELESS:
-        case DXGI_FORMAT_BC7_UNORM:
-        case DXGI_FORMAT_BC7_UNORM_SRGB:
-            return 8;
-
-#if defined(_XBOX_ONE) && defined(_TITLE)
-
-        case DXGI_FORMAT_R10G10B10_7E3_A2_FLOAT:
-        case DXGI_FORMAT_R10G10B10_6E4_A2_FLOAT:
-        case DXGI_FORMAT_R10G10B10_SNORM_A2_UNORM:
-            return 32;
-
-        case DXGI_FORMAT_D16_UNORM_S8_UINT:
-        case DXGI_FORMAT_R16_UNORM_X8_TYPELESS:
-        case DXGI_FORMAT_X16_TYPELESS_G8_UINT:
-            return 24;
-
-        case DXGI_FORMAT_R4G4_UNORM:
-            return 8;
-
-#endif // _XBOX_ONE && _TITLE
-
-        default:
-            return 0;
-        }
-    }
-
-    //--------------------------------------------------------------------------------------
-    // Determines if the format is block compressed
-    //--------------------------------------------------------------------------------------
-    bool IsCompressed(_In_ DXGI_FORMAT fmt)
-    {
-        switch (fmt)
-        {
-        case DXGI_FORMAT_BC1_TYPELESS:
-        case DXGI_FORMAT_BC1_UNORM:
-        case DXGI_FORMAT_BC1_UNORM_SRGB:
-        case DXGI_FORMAT_BC2_TYPELESS:
-        case DXGI_FORMAT_BC2_UNORM:
-        case DXGI_FORMAT_BC2_UNORM_SRGB:
-        case DXGI_FORMAT_BC3_TYPELESS:
-        case DXGI_FORMAT_BC3_UNORM:
-        case DXGI_FORMAT_BC3_UNORM_SRGB:
-        case DXGI_FORMAT_BC4_TYPELESS:
-        case DXGI_FORMAT_BC4_UNORM:
-        case DXGI_FORMAT_BC4_SNORM:
-        case DXGI_FORMAT_BC5_TYPELESS:
-        case DXGI_FORMAT_BC5_UNORM:
-        case DXGI_FORMAT_BC5_SNORM:
-        case DXGI_FORMAT_BC6H_TYPELESS:
-        case DXGI_FORMAT_BC6H_UF16:
-        case DXGI_FORMAT_BC6H_SF16:
-        case DXGI_FORMAT_BC7_TYPELESS:
-        case DXGI_FORMAT_BC7_UNORM:
-        case DXGI_FORMAT_BC7_UNORM_SRGB:
-            return true;
-
-        default:
-            return false;
-        }
-    }
-
-    //--------------------------------------------------------------------------------------
-    // Get surface information for a particular format
-    //--------------------------------------------------------------------------------------
-    void GetSurfaceInfo(_In_ size_t width,
-        _In_ size_t height,
-        _In_ DXGI_FORMAT fmt,
-        _Out_opt_ size_t* outNumBytes,
-        _Out_opt_ size_t* outRowBytes,
-        _Out_opt_ size_t* outNumRows)
-    {
-        size_t numBytes = 0;
-        size_t rowBytes = 0;
-        size_t numRows = 0;
-
-        bool bc = false;
-        bool packed = false;
-        bool planar = false;
-        size_t bpe = 0;
-        switch (fmt)
-        {
-        case DXGI_FORMAT_BC1_TYPELESS:
-        case DXGI_FORMAT_BC1_UNORM:
-        case DXGI_FORMAT_BC1_UNORM_SRGB:
-        case DXGI_FORMAT_BC4_TYPELESS:
-        case DXGI_FORMAT_BC4_UNORM:
-        case DXGI_FORMAT_BC4_SNORM:
-            bc = true;
-            bpe = 8;
-            break;
-
-        case DXGI_FORMAT_BC2_TYPELESS:
-        case DXGI_FORMAT_BC2_UNORM:
-        case DXGI_FORMAT_BC2_UNORM_SRGB:
-        case DXGI_FORMAT_BC3_TYPELESS:
-        case DXGI_FORMAT_BC3_UNORM:
-        case DXGI_FORMAT_BC3_UNORM_SRGB:
-        case DXGI_FORMAT_BC5_TYPELESS:
-        case DXGI_FORMAT_BC5_UNORM:
-        case DXGI_FORMAT_BC5_SNORM:
-        case DXGI_FORMAT_BC6H_TYPELESS:
-        case DXGI_FORMAT_BC6H_UF16:
-        case DXGI_FORMAT_BC6H_SF16:
-        case DXGI_FORMAT_BC7_TYPELESS:
-        case DXGI_FORMAT_BC7_UNORM:
-        case DXGI_FORMAT_BC7_UNORM_SRGB:
-            bc = true;
-            bpe = 16;
-            break;
-
-        case DXGI_FORMAT_R8G8_B8G8_UNORM:
-        case DXGI_FORMAT_G8R8_G8B8_UNORM:
-        case DXGI_FORMAT_YUY2:
-            packed = true;
-            bpe = 4;
-            break;
-
-        case DXGI_FORMAT_Y210:
-        case DXGI_FORMAT_Y216:
-            packed = true;
-            bpe = 8;
-            break;
-
-        case DXGI_FORMAT_NV12:
-        case DXGI_FORMAT_420_OPAQUE:
-            planar = true;
-            bpe = 2;
-            break;
-
-        case DXGI_FORMAT_P010:
-        case DXGI_FORMAT_P016:
-            planar = true;
-            bpe = 4;
-            break;
-
-#if defined(_XBOX_ONE) && defined(_TITLE)
-
-        case DXGI_FORMAT_D16_UNORM_S8_UINT:
-        case DXGI_FORMAT_R16_UNORM_X8_TYPELESS:
-        case DXGI_FORMAT_X16_TYPELESS_G8_UINT:
-            planar = true;
-            bpe = 4;
-            break;
-
-#endif
-        }
-
-        if (bc)
-        {
-            size_t numBlocksWide = 0;
-            if (width > 0)
-            {
-                numBlocksWide = std::max<size_t>(1, (width + 3) / 4);
-            }
-            size_t numBlocksHigh = 0;
-            if (height > 0)
-            {
-                numBlocksHigh = std::max<size_t>(1, (height + 3) / 4);
-            }
-            rowBytes = numBlocksWide * bpe;
-            numRows = numBlocksHigh;
-            numBytes = rowBytes * numBlocksHigh;
-        }
-        else if (packed)
-        {
-            rowBytes = ((width + 1) >> 1) * bpe;
-            numRows = height;
-            numBytes = rowBytes * height;
-        }
-        else if (fmt == DXGI_FORMAT_NV11)
-        {
-            rowBytes = ((width + 3) >> 2) * 4;
-            numRows = height * 2; // Direct3D makes this simplifying assumption, although it is larger than the 4:1:1 data
-            numBytes = rowBytes * numRows;
-        }
-        else if (planar)
-        {
-            rowBytes = ((width + 1) >> 1) * bpe;
-            numBytes = (rowBytes * height) + ((rowBytes * height + 1) >> 1);
-            numRows = height + ((height + 1) >> 1);
-        }
-        else
-        {
-            size_t bpp = BitsPerPixel(fmt);
-            rowBytes = (width * bpp + 7) / 8; // round up to nearest byte
-            numRows = height;
-            numBytes = rowBytes * height;
-        }
-
-        if (outNumBytes)
-        {
-            *outNumBytes = numBytes;
-        }
-        if (outRowBytes)
-        {
-            *outRowBytes = rowBytes;
-        }
-        if (outNumRows)
-        {
-            *outNumRows = numRows;
-        }
-    }
-
-    //--------------------------------------------------------------------------------------
-    DXGI_FORMAT EnsureNotTypeless(DXGI_FORMAT fmt)
-    {
-        // Assumes UNORM or FLOAT; doesn't use UINT or SINT
-        switch (fmt)
-        {
-        case DXGI_FORMAT_R32G32B32A32_TYPELESS: return DXGI_FORMAT_R32G32B32A32_FLOAT;
-        case DXGI_FORMAT_R32G32B32_TYPELESS:    return DXGI_FORMAT_R32G32B32_FLOAT;
-        case DXGI_FORMAT_R16G16B16A16_TYPELESS: return DXGI_FORMAT_R16G16B16A16_UNORM;
-        case DXGI_FORMAT_R32G32_TYPELESS:       return DXGI_FORMAT_R32G32_FLOAT;
-        case DXGI_FORMAT_R10G10B10A2_TYPELESS:  return DXGI_FORMAT_R10G10B10A2_UNORM;
-        case DXGI_FORMAT_R8G8B8A8_TYPELESS:     return DXGI_FORMAT_R8G8B8A8_UNORM;
-        case DXGI_FORMAT_R16G16_TYPELESS:       return DXGI_FORMAT_R16G16_UNORM;
-        case DXGI_FORMAT_R32_TYPELESS:          return DXGI_FORMAT_R32_FLOAT;
-        case DXGI_FORMAT_R8G8_TYPELESS:         return DXGI_FORMAT_R8G8_UNORM;
-        case DXGI_FORMAT_R16_TYPELESS:          return DXGI_FORMAT_R16_UNORM;
-        case DXGI_FORMAT_R8_TYPELESS:           return DXGI_FORMAT_R8_UNORM;
-        case DXGI_FORMAT_BC1_TYPELESS:          return DXGI_FORMAT_BC1_UNORM;
-        case DXGI_FORMAT_BC2_TYPELESS:          return DXGI_FORMAT_BC2_UNORM;
-        case DXGI_FORMAT_BC3_TYPELESS:          return DXGI_FORMAT_BC3_UNORM;
-        case DXGI_FORMAT_BC4_TYPELESS:          return DXGI_FORMAT_BC4_UNORM;
-        case DXGI_FORMAT_BC5_TYPELESS:          return DXGI_FORMAT_BC5_UNORM;
-        case DXGI_FORMAT_B8G8R8A8_TYPELESS:     return DXGI_FORMAT_B8G8R8A8_UNORM;
-        case DXGI_FORMAT_B8G8R8X8_TYPELESS:     return DXGI_FORMAT_B8G8R8X8_UNORM;
-        case DXGI_FORMAT_BC7_TYPELESS:          return DXGI_FORMAT_BC7_UNORM;
-        default:                                return fmt;
-        }
-    }
-
     //--------------------------------------------------------------------------------------
     HRESULT CaptureTexture(_In_ ID3D11DeviceContext* pContext,
         _In_ ID3D11Resource* pSource,
