+++ Src/SharedResourcePool.h
@@ -25,7 +25,7 @@ namespace DirectX
     // This is used to avoid duplicate resource creation, so that for instance a caller can
     // create any number of SpriteBatch instances, but these can internally share shaders and
     // vertex buffer if more than one SpriteBatch uses the same underlying D3D device.
-    template<typename TKey, typename TData>
+    template<typename TKey, typename TData, typename... TConstructorArgs>
     class SharedResourcePool
     {
     public:
@@ -37,7 +37,7 @@ namespace DirectX
         SharedResourcePool& operator= (SharedResourcePool const&) = delete;
 
         // Allocates or looks up the shared TData instance for the specified key.
-        std::shared_ptr<TData> DemandCreate(TKey key)
+        std::shared_ptr<TData> DemandCreate(TKey key, TConstructorArgs... args)
         {
             std::lock_guard<std::mutex> lock(mResourceMap->mutex);
 
@@ -55,7 +55,7 @@ namespace DirectX
             }
             
             // Allocate a new instance.
-            auto newValue = std::make_shared<WrappedData>(key, mResourceMap);
+            auto newValue = std::make_shared<WrappedData>(key, mResourceMap, args...);
 
             mResourceMap->insert(std::make_pair(key, newValue));
 
@@ -77,10 +77,10 @@ namespace DirectX
         // to remove instances from our pool before they are freed.
         struct WrappedData : public TData
         {
-            WrappedData(TKey key, std::shared_ptr<ResourceMap> const& resourceMap)
+            WrappedData(TKey key, std::shared_ptr<ResourceMap> const& resourceMap, TConstructorArgs... args)
               : mKey(key),
                 mResourceMap(resourceMap),
-                TData(key)
+                TData(key, args...)
             { }
 
             ~WrappedData()
