+++ Src/WICTextureLoader.cpp
@@ -30,12 +30,11 @@
 
 #include "pch.h"
 
-#include <wincodec.h>
-
 #include "WICTextureLoader.h"
 
 #include "DirectXHelpers.h"
 #include "PlatformHelpers.h"
+#include "LoaderHelpers.h"
 
 using namespace DirectX;
 using Microsoft::WRL::ComPtr;
@@ -263,37 +262,6 @@ namespace
             return 0;
 
         return bpp;
-    }
-
-    //--------------------------------------------------------------------------------------
-    DXGI_FORMAT MakeSRGB(_In_ DXGI_FORMAT format)
-    {
-        switch (format)
-        {
-        case DXGI_FORMAT_R8G8B8A8_UNORM:
-            return DXGI_FORMAT_R8G8B8A8_UNORM_SRGB;
-
-        case DXGI_FORMAT_BC1_UNORM:
-            return DXGI_FORMAT_BC1_UNORM_SRGB;
-
-        case DXGI_FORMAT_BC2_UNORM:
-            return DXGI_FORMAT_BC2_UNORM_SRGB;
-
-        case DXGI_FORMAT_BC3_UNORM:
-            return DXGI_FORMAT_BC3_UNORM_SRGB;
-
-        case DXGI_FORMAT_B8G8R8A8_UNORM:
-            return DXGI_FORMAT_B8G8R8A8_UNORM_SRGB;
-
-        case DXGI_FORMAT_B8G8R8X8_UNORM:
-            return DXGI_FORMAT_B8G8R8X8_UNORM_SRGB;
-
-        case DXGI_FORMAT_BC7_UNORM:
-            return DXGI_FORMAT_BC7_UNORM_SRGB;
-
-        default:
-            return format;
-        }
     }
 
     //---------------------------------------------------------------------------------
