+++ Src/dds.h
@@ -21,14 +21,6 @@
 
 #pragma once
 
-#if defined(_XBOX_ONE) && defined(_TITLE)
-#include <d3d11_x.h>
-#else
-#include <dxgiformat.h>
-#endif
-
-#include <stdint.h>
-
 
 namespace DirectX
 {
@@ -233,4 +225,4 @@ struct DDS_HEADER_DXT10
 static_assert( sizeof(DDS_HEADER) == 124, "DDS Header size mismatch" );
 static_assert( sizeof(DDS_HEADER_DXT10) == 20, "DDS DX10 Extended Header size mismatch");
 
-}; // namespace
+} // namespace
